TestPlatform/TestPlatform/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
colorlog==4.2.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.8
django-cors-headers==3.4.0
django-filter==2.3.0
django-rest-swagger==2.2.0
django-restframework==0.0.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1
har2case==0.3.1
HttpRunner==1.5.8
idna==2.10
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==2.0.1
openapi-codec==1.3.2
packaging==20.4
PyJWT==1.7.1
PyMySQL==0.10.0
pyparsing==2.4.7
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-toolbelt==0.9.1
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
